use cstdio for test output
